+++ openair1/PHY/defs_RU.h
@@ -132,7 +132,7 @@ typedef struct {
   /// \brief Anaglogue beam ID for each OFDM symbol (used when beamforming not done in RU)
   /// - first index: antenna port
   /// - second index: beam_id [0.. symbols_per_frame[
-  uint8_t **beam_id;
+  int **beam_id;
 } RU_COMMON;
